refactor: club, direct, chatroom repository 계층 패키지 변경
