+++ src/main/java/com/mobble/mobbleserver/application/chat/room/service/common/ChatRoomModifyService.java
@@ -7,9 +7,9 @@ import com.mobble.mobbleserver.domain.chat.room.ChatRoom;
 import com.mobble.mobbleserver.domain.chat.room.ChatRoomType;
 import com.mobble.mobbleserver.domain.member.Member;
 import com.mobble.mobbleserver.refactor.chat.chatMessage.repository.ChatMessageRepository;
-import com.mobble.mobbleserver.refactor.chat.chatRoom.repository.ChatRoomRepository;
+import com.mobble.mobbleserver.infrastructure.persistence.chat.common.ChatRoomRepository;
 import com.mobble.mobbleserver.refactor.chat.chatRoom.validator.ChatRoomValidator;
-import com.mobble.mobbleserver.refactor.chat.chatRoomParticipant.repository.ChatRoomParticipantRepository;
+import com.mobble.mobbleserver.infrastructure.persistence.chat.common.ChatRoomParticipantRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
