+++ src/main/java/com/mobble/mobbleserver/infrastructure/persistence/chat/club/ClubRoomInfoRepository.java
@@ -1,0 +1,9 @@
+package com.mobble.mobbleserver.infrastructure.persistence.chat.club;
+
+import com.mobble.mobbleserver.domain.chat.room.ClubRoomInfo;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ClubRoomInfoRepository extends JpaRepository<ClubRoomInfo, Long> {
+
+    boolean existsClubRoomInfoByClubId(Long clubId);
+}
