+++ src/main/java/com/mobble/mobbleserver/infrastructure/persistence/chat/direct/DirectChatRoomRepository.java
@@ -1,4 +1,4 @@
-package com.mobble.mobbleserver.refactor.chat.directChatRoom.repository;
+package com.mobble.mobbleserver.infrastructure.persistence.chat.direct;
 
 import com.mobble.mobbleserver.domain.chat.room.DirectRoomInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
